Title: 실수 받아 그대로 출력, Time: 145ms, Memory: 11MB, Status: Wrong Answer - CodeTree

+++ 250211/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
 
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextDouble();
+        double a = sc.nextDouble();
         sc.close();
         System.out.print(a + 0.58);
     }
